Renamed install folder to scripts

+++ scripts/links.sh
@@ -1,4 +1,10 @@
 #! /bin/bash
+
+if [ "${PWD##*/}" != "dotfiles" ]; then
+    echo "[ WARN ] Run this script from inside the dotfiles directory"
+    exit
+fi
+
 echo "[ INFO ] Creating symlink"
 rm ~/.zshrc
 ln -s $(pwd)/.zshrc ~/.zshrc
